Oh no! Yes, we have been! 😱
Looking at the code, we'\''re setting `zoom={0}` for 1×, which should be fine, BUT - `expo-camera`'\''s CameraView is using the **wide-angle camera by default**, not the main camera!

On iPhone, there are typically:
- **0.5×** - Ultra-wide lens
- **1×** - Main/wide lens (the best quality one)
- **2×-5×** - Telephoto lenses

The issue is that CameraView doesn'\''t let us select which physical lens to use - it just gives us a zoom slider on whatever the default camera is.

Unfortunately, **expo-camera'\''s CameraView doesn'\''t support lens selection**. The old Camera API had it, but they simplified CameraView and removed that feature. This is a known limitation.

**Here'\''s the reality:**
- We CAN'\''T access the actual telephoto lenses with expo-camera
- We'\''re stuck with digital zoom only
- Your iPhone'\''s beautiful optical zoom is unavailable 😢

**Our options:**

1. **Remove the zoom feature entirely** - Just use the default camera
2. **Switch to `