I need to clarify what you'\''re asking for. There are two possible interpretations:
**Option 1: Allow YouTube/music to keep playing when user opens PanHandler app**
- Currently, when users open your app, it might stop their background music/YouTube
- We'\''d configure the app to NOT interrupt audio from other apps
- Users can measure while listening to music/podcasts

**Option 2: Embed YouTube video player inside PanHandler app**
- Add a feature to play YouTube videos within the app
- Keep videos playing in background while measuring
- More complex, requires YouTube API integration

---

Which one do you want?

My guess is **Option 1** - you want users to be able to listen to music/YouTube while using PanHandler for measurements, right?

If that'\''s the case, I can configure the audio session to be non-intrusive so the app doesn'\''t stop background audio when it launches.