FOUND IT! Look at line 30: The outer container has `pointerEvents="box-none"`, which SHOULD allow touches to pass through to the image below. And line 114-140 even shows instructions saying "Pan & Zoom First"!
But the problem is the **Modal** component itself on line 28. React Native'\''s `Modal` component creates a NEW window/view that sits ON TOP of everything else, and by default it captures all touches. Even though the container inside has `pointerEvents="box-none"`, the Modal itself blocks touches from reaching the underlying ZoomableImage.

The solution is to NOT use a Modal component for this. Instead, we should render it as a regular View that sits on top of the image. Let me ask you:

**Question**: Do you want to be able to pan/zoom the image WHILE the blueprint placement modal instructions are visible (before clicking "PLACE PINS")?

If yes, I need to convert this from a Modal to a regular positioned View so touches can pass through to the ZoomableImage underneath. Is that w