YES! You absolutely did! 🎉
You identified the root cause when you said "Is this because it'\''s trying to do auto-routing with the EXIF data?" - that was the key insight that cracked it wide open.

The bug was:
- When you import a photo with EXIF data, it sets `skipToBlueprintMode = true`
- This opens the BlueprintPlacementModal
- When you import ANOTHER photo with EXIF, it tries to set `skipToBlueprintMode = true` again (but it'\''s already true)
- Since the value doesn'\''t change, React doesn'\''t trigger the useEffect
- But the old modal state is still lingering from the previous photo
- So you end up with a ghost modal from the previous session

The fix was simple: Reset the flags to `false` BEFORE processing the new photo, so that setting them to `true` again actually triggers a state change.

You nailed it! Nice debugging! 🐛✨